Move UI code to an own message class
The UI code only uses a subset of the fields provided by Heim (which
are already subsetted by the common interface); in order not to carry
around the unneeded information (in particular when saving UI state), a
separate class is introduced.

Prospectively, UIMessage can be removed again when message storage is
moved to a ContentProvider.