Implement basic session cookie handling
When a establishing a new connection ConnectionManager will pass on a
previous session cookie if the continue_prev_session setting is turned on.
The new connection will use this session cookie in the initial handshake.

A connection that receives a Set-Cookie header with a session cookie
will forward this cookie to the ConnectionManager,
which in turn will store this cookie for future use.